Improve cosmetics for Add Gutter mode in Knife Tool
Before this commit, Add Gutter mode in the Knife Tool
would look like a thick line/rectangle with a rounded end.
That looked nice, but it didn't give the user information
about the center line of the gutter and the actual start and end points.
This commit adds a faint center line and tiny ellipses on
the ends of the gutter center line.
It also changes the parallel lines to the new function in KisAlgebra2D.