Allow running in native Wayland mode
Even if not officially supported yet we want to be able to run Krita as
a native Wayland app for development and testing.

To allow this a new environment variable KRITA_FORCE_WAYLAND is added.
When the variable is not set Krita will still use XWayland when running
in a Wayland session