UI/Animation/TimelineDocker: Can now re-select previous selected frame after pausing.
The default behavior of QAbstractItemModel was causing an edge case
where users couldn't re-select the starting frame of their animation after pausing
because it was technically already paused. Now, we've made it so that pausing
the animation playback will move the selection to the frame that the
animation was paused on, solving this issue.

BUG:500885