Add the Knife Tool icon to the qrc file
Before this commit, the Knife Tool/Comic Panel Editing Tool had
the icon in the code, but it wasn't added
to the repository or to the qrc file so it couldn't be used.
This commit adds it to the qrc file and repo so the Toolbox finally has
an icon for the tool.