Stop QImageReader from asserting when creating large thumbnails for KRA
I hit this while Krita is creating thumbnails for some of my larger KRA
files. To be more specific, this is because it picks "mergedimage.png"
which is quite large instead of the smaller res "preview.png" which
seems to be intended.

Anyway, this is probably better to disable for now to match Qt5 behavior
and re-enable later down the line.