Import and use QtGuiPrivate instead of using an implementation variable
While this worked in Qt5 and some of Qt6, in the latest Qt6 dev they
have been shuffling their private CMake modules just enough to break
this.

This isn't how we're supposed to use the private headers anyway, it's
much better (and more reliable) to find QtGuiPrivate and link to it. For
some reason Qt5 makes it more difficult to get to Qt5GuiPrivate (hence
why it's in a separate find_package call.)