TenBrushes: Add an option to disable on-canvas messages
When activating brush presets, the ten brushes plugin displays an
on-canvas message indicating which one you activated. Sometimes users do
not desire this, or they only use one or two presets. For example, I
only have an Eraser preset - the on-screen message only ever displays
that and my previous brush.

This adds a setting to disable said message, and it's still left on by
default.