Optimize fetching metadata for resources
When doing filtering in item delegates we don't need to fetch
the entire set of all the metadata available in the resource.
Instead, we should fetch only information we need. Using a
query with a limit. It is much more efficient.

The patch should also be coupled with an index in the database.