Add Marker blend mode
This mode is like Alpha Darken, except it adheres to channel flags and
interpolates colors without artifacting by using `srcAlpha / dstAlpha`
instead of just `srcAlpha` as the interpolation parameter. When you use
it on a brush with the Build-Up paint mode it works like SAI's marker
tool, only increasing but not compounding opacity and mixing the colors.

The implementation splits off the alpha calculation in alpha darken into
a separate function so that the marker mode can make use of it without
intruding unto the original code too much.