Fix initialization of the global instance of KisStorageModel
When KisResourceCacheDb::synchronizeStorage() performs
synchronization of the resources, some of the resources being
oaded may ask for the global KisStorageModel, which will
make it created.

When all the storages are synchronized, we should notify the
model about the synchronization completed, so that it could
reread the list of storages from the database.