Import and use QtGuiPrivate instead of using an implementation variable
Previously Krita dependend on an implementation detail of Qt's CMake
module, which has a history of breaking. For reference, Krita no longer
builds on Qt dev because they shuffled things around again!

This isn't how we're supposed to use the private headers anyway, it's
more reliable to use the QtGuiPrivate CMake module. Unfortunately, not
even *this* is consistent and there's three slightly different ways to
access it across Qt5, Qt6 < 6.10 and Qt6 >= 6.10. All of them should be
covered now.