Fix initialization of the foreign_keys constraint for DB
It turns out that when writing resources support we
accidentially forgot to enable the foreign keys constraint.
So now we need to enable it in iterative way:

1) All unstable (beta, release_candidate) and nigtly builds will have
    the foreign_keys **enabled**.

2) All stable released versions will have the foreign_keys **disabled**.

This way we should find and track down all the issues that could be caused
by this (quite significant) change.

PS:
The patch also removes a huge part of kritaversion.h, that was supposed to
change Krita's version to X.Y.99 in case of unstable. We never used this feature
and **it was broken** (KRITA_STABLE definition was defined **after**
the detection code)