Fix a crash when autosaving extremely heavy images
If the image

1) was too heavy making the autosave slow
2) had a lot of compositions making KisImage copy-ctor slow
3) had an adjustment layer with vector selection

when all the three conditions were met, Krita could crash, because
KisShapeSelectionModel inside the adjustment layer could emit an update
signal, which would cause the image pointer to be upgraded to a **strong**
shared pointer. If the timing was unfortune enough for the copy-ctor not to be
finished yet, it could cause the copied image to be destroyed right in
the middle of the copy operation.

To avoid that, this patch sets teh resolution value **before** the layers are
assigned to a new image. It makes sure that no signals are emitted in the
meantime.