Rework KoTextShape::outline so it returns a single shape.
This is both more useful because a single parent groupShape can have its children
inherit, and in the cases there's only a single childShape, we return that instead,
which should make it less confusing to edit (because groupShapes are only disintuishable
by whether they have multiple shapes, a single shape inside a groupshape currently looks
the same as the same shape, outside the group. This resulted in confusion for artists
trying to use the 'convert to Shape' command for the new text shape).