Rewrite KoPluginLoader::load()
...and cover the whole KoPluginLoader with unittests.

The great bunch of code  in KoPluginLoader::load() was basically dead
and didn't work, because of API changes that happened in the meantime.
The patch covers it fully with a unittest and rewrites against the new
KPlugin metadata standards.

KoPluginLoaderTest basically builds a few dummy plugins and then tries
to load them in various ways. These dummy plugins should never be installed
into the final installation root directory.

The patch also replaces all the usages of 'X-Flake-PluginVersion' in favour of
'X-Krita-Version', since the latter variant is newer. We should probably consider
using 'X-KDE-PluginInfo-Version' instead, though I'm not very sure.