Fix synchronization of the Qt's rendering pipeline with the one in Krita
When we deletage updates compression to Qt, we should still handle
the case when our own openGL context has not completed execution
of the canvas rendering code. In which case we should just resent one
more update() call, which will be queued in QWidgetRepaintManager
and processed during the next frame cycle.