BUGFIX: Fixed incorrect scaling of animated transform mask values.
It seems that when scaling a canvas that contains an
animated transform mask, the transform parameter scaling
function was creating new keyframes out of the previously
interpolated in-between values, which was causing
the transform mask to act in an unpredictable and
undesirable way.

Now we simply scale the scalar keyframes as needed.

BUG:469881
CCBUG:475336