BUGFIX: Static opacity changes now properly clear animation cache.
Note: This fix is referring to the changing of the general layer opacity of
a RASTER animated layer, NOT a layer with scalar/curves opacity animation.

This bug fixes an issue where if the user changes the overall layer
opacity of a layer with multiple drawn frames of RASTER animation,
the cache would only be invalidated for 1 of the frames, despite non-animated
layer opacity being something that should apply to all frames equally.

Now chaging the opacity will invalidate the entire animation cache, so that
layer opacity changes will be immediately reflected as expected.

HACK: As noted in the comment, this fix is something of a hack, as the frames
are being invalidated directly within the KisNodeOpacityCommand. A better
fix would be to slightly refactor KisNode::setDirty, KisImage::requestProjectionUpdate,
and/or KisProjectionUpdateFlags, to handle more variations of cache invalidation
behavior.

BUG:499389