Implement option for ignoring of F13-F24 keys on Windows
Some apps (e.g. WeeChat) uses emulated F22 key press to bring
itself to the toplevel. Since the key is unbalanced, it confuses Krita.
So we implemented an option just to ignore all F13+ keys for users
who have this problem.

Thanks Wendy Gan (@gwdx) for the original suggestion of the fix in
https://invent.kde.org/graphics/krita/-/merge_requests/2416. This patch
also implements this change, but makes it configurable.

BUG:483687