Speed up the font family dropdown in the text properties docker.
The main issue was that the ListView in the ResourceView was instantiated
as full height, which meant that even as the ListView was reduced to a smaller
size later on, it had already laid out every single delegate it could possibly
need, which mean every single font was being drawn.

Beyond that, the KoShapeQtQuickLabel was wrapped into a Loader, and Reuse items
was turned on for OpenType and FontFamilies, so that these complex delegates
are not being instantiated all the time.