Index Colors: fix crash when filtering non-transparent layer
If there is just one color, the diffs array will never have
a value bigger than std::numeric_limits<float>::min(). That's
not a bug, though, the color we return should in that case be
the single color that is present in the layer.

This does not reintroduce BUG:473242 which was fixed in
74a8cf5472249d05a8a4dd9350e3e2331db3c58c , but does restore
part of the original meaning of the original author (who
also made primaryColor 0.