Fix slowdown when resizing Krita
When KCatalog asks gettext about the translation
of the string it passes some information to that via
LANGUAGE environment variable. If KCatalog sees
that LANGUAGE is not consistent with the current
language selection, then it sets it manually... **for
every single translatable string**. And it seems like
qputenv() is a bit slow on Windows. Which causes
KisLayer::sectionModelProperties() to become glowingly
hot on resizing Krita window.

So this patch just makes sure that LANGUAGE
environment variable is initialized correctly on Krita
start to avoid this manual hot spot.

PS:
Yes, we should also move layer's properties into some static storage, but that is a different topic, because there are other
places that are both translatable and re-renderable on window
resize.