Make it so that the Style Preset sample svg can have inherited color.
This ensures that artists can tell whether a preset sets the color at all.
For this, inheritance needs to be on by default, but because I cannot oversee
the consequences of that change, this is encapsulated with a boolean in the
parser. There's still some questions to be had about what happens if the
style is explicitely black, but we'll cross that bridge when we can configure
the color on the text explcitely.