Add "Include Current Layer" option when referencing color labels on "fill" tool and "enclose and fill" tool
This merge request adds the ability to include the current layer as a reference layer when filling using color label regions in the fill tool and the enclose and fill tool.

A checkbox has been added to the color label selector widget from within the tool options when either of the tools are selected:

![2025-07-21_20-18](/uploads/cab459b1f0e3e4ae522682bfacb8c5f2/2025-07-21_20-18.png)

When the checkbox is selected, the layer that is currently being edited will be included with layers that match the appropriate color label that is selected.

I wanted to add this because it can be cumbersome to have to temporarily add a layer to a color label when you are only working on it for a moment while drawing. While you can also reference no color labels, this can include extraneous lines that you might not want to reference by.

Test Plan
---------
To test, create two layers, one with a blue color label and one with no layer. 

Create a fillable shape on both of the layers.

When "Include current layer" is selected, you will be able to fill the current layer if and only if it is selected.

Formalities Checklist
--------------------- 

- [x] I confirmed this builds.
- [x] I confirmed Krita ran and the relevant functions work.
- [x] I tested the relevant unit tests and can confirm they are not broken. (If not possible, don't hesitate to ask for help!)
- [x] I made sure my commits build individually and have good descriptions as per [KDE guidelines](https://community.kde.org/Policies/Commit_Policy).
- [x] I made sure my code conforms to the standards set in the HACKING file.
- [x] I can confirm the code is licensed and attributed appropriately, and that unattributed code is mine, as per [KDE Licensing Policy](https://community.kde.org/Policies/Licensing_Policy).
- [ ] Does the patch add a user-visible feature? If yes, is there a documentation MR ready for it at [Krita Documentation Repository](https://invent.kde.org/documentation/docs-krita-org)?

_**Reminder: the reviewer is responsible for merging the patch, this is to ensure at the least two people can build the patch. In case a patch breaks the build, both the author and the reviewer should be contacted to fix the build.**_
_**If this is not possible, the commits shall be reverted, and a notification with the reasoning and any relevant logs shall be sent to the mailing list, kimageshop@kde.org.**_