Add two new custom toolbars to Krita
Before this commit, there were only two, maybe three (including
editToolBar which is commented out right at the time of writing)
toolbars in Krita. This commit adds two custom ones
that the users can use for vertical toolbars or anything else.
One toolbar is intended to be a vertical docker toolbar next to
the toolbox, but the user can use it as something else
if they prefer. I added an additional one for additional
customizeability.