Fix drifiting of the canvas when zooming and rotating
The problem is that now the still points in the canvas transformations
cannot be satisfied precisely, since the canvas is always aligned to the
hardware pixel grid. This causes severe drifts when zooming and rotating
the canvas using the mouse/stylus gestures.

To fix the issue, this patch changes the concept of a "still point". Now a
still point is a pair of both, image and widget points that correspond to
each other. It means that if we keep the "still point" unchanged during
the entire zoom or rotation operation, then the image will not drift
anywhere, even though the still point is not kept "still" for the duration
of the operation.

The patch also introduces a concept of preferredTransformationCenter(),
which is basically a point on the image, that is currently mapped to the
center of the widget. This "preferred center" as a default center for such
operations as Ctrl+"+" and Ctrl+"-" shortcuts.

And lastly, the patch disables pixel alignment for the canvas when it is
rotates to any angle that is "not orthogonal" (i.e. not multiple of 90 deg).
Otherwise rotation looks a bit dizzy (even when doesn't drift anywhere).