Fix double resize of the canvas widget on resetting the canvas
This patch fixes an assert when the user changes the canvas type
from QPainter to openGL one. The assert was caused by the widget
to be resized twice: once with a random size and then with a proper
size set in resetLayout().

The patch fixes that by resizing the canvas in invisible state.