Add android:windowOptOutEdgeToEdgeEnforcement for Android-35
android-35 API forces all applications to respect edge-to-edge
feature of Android. In this mode, the application window is
stretched throughout the entire screen, while the system notifies
where the OS-specific controls are situated using so-called "insets".

In this scenario the application should handle this insets and make
sure that no useful controls are placed under these "insets".

It makes the screen to look a little bit prettier, when the system
controls are translucent.

Qt5 doesn't support handling of these insets. They are only handled
in Qt 6.9.0:

https://codereview.qt-project.org/c/qt/qtbase/+/608508

I have tried to backport this patch, but it seems to be impossible,
becasue of too many changes and refactorings happened in the meantime.

So for now we just use a special killswitch that adds an exemption
for the app from the edge-to-edge feature. This exemption will work
only for android-35. In android-36 it will not be available, so we
will be forced to update to Qt 6.9.0.