Add Keep Aspect Ratio action to Transform Tool context menu
Before this commit, the only way to set the keep aspect ratio setting
(which is pretty important) without a keyboard was by switching to Scale
options, then scrolling down and sometimes scrolling sideways to get
to the aspect ratio button.
This commit adds an action in the context menu (right-click) that does
the same thing.

This commit is related to MR !2123. It doesn't completely fix the issue
of Keep Aspect Ratio on Android, but it does make it slightly easier to
access.
For full fix, a new design is required. It should be an option
easily available in the options widget, but where?
And it might or might not be either always on by default (on every
tool activation) OR it should be saved from the last use
of the transform tool.