UX/Animation: Switch to active KEYFRAME when inserting/removing hold frames.
When inserting or removing hold frames, we now switch the animation time
to the time of the currently active keyframe.

This is meant to make it easier to continue drawing on the correct frame
when adjusting hold frames could otherwise cause the keyframe you're
drawing on to change position from under you.

(Follow up work to previous hold frames bug.)

CCBUG:507380