Fix KoZoomAction and test_animated_transform_parameters Test.
This fixes TestKoZoomAction and test_animated_transform_parameters.

TestCompositeOpInversion and kis_strokes_queue_test just pass on my system. I think with the former the problem is that it is just too big and takes too long, so the CI thinks it froze. For the latter, I have no clue.