Split the canvas observer from the text properties docker into a qml object.
By separating this and instantiating it inside QML reduces the amount of
calls we need to do onto the QQuickWidget rootObject, reducing in turn the
complexity of the signal handing.

With this we are also making the TextProperties themselves a internal property
instead of a context-property, which allows us to test against updates for a
single subproperty, which in turn we don't need to do spurious updates whenever
the text properties get updated at all. The axesmodel and font features in
particular were troubled by this.

There's still some minor issues, like the units not converting, though I am
unsure if that weren't already broken.