Introduce hasFocus in the textproperties canvas observer.
There was a situation where, when switching text shapes, one shapes' properties
got set to the other. This was because the properties were stored inside the
docker and the compressor was still running. To avoid this, we check if the
text properties docker has focus before deciding whether we want to accept
the properties from the canvas provider (and thus stay up-to-date with the
selected text shape).