[android] Properly synchronize saver service
The document saver service is lacking proper synchronization with
itself, which could cause various kinds of mayhem. Long save operations
could cause two saver threads to be spawned, which would un-foreground
each other. Swiping closing the application while it was in the
background would attempt to foreground the service just so it could
terminate the currently running foreground service. And there was no
proper bookkeeping of the data structures, which could potentially cause
races.

This commit now properly synchronizes the access to the data structures
and only foregrounds the service when a save is started. When a save
request is received while another one is already running, it will
instead just set a flag that will make the save thread run again instead
of starting a second instance of it. Terminating the application or
cancelling the save will no longer attempt to create a foreground thread
for the purpose of terminating it.

Relates to #15.