[android] Catch link error in save service
Since the document saver service runs in a background thread, you can
get sufficiently unlucky with losing the race and end up with it
starting its work when the application's native library has already been
unloaded. This leads to an UnsatisfiedLinkError and a crash.

This commit just catches that error, since there's not much the service
can do when the application is gone.

Relates to #15.