Overhaul the font resource dropdown so it has a textinput.
For some people this might be a more intuitive way to interact
with the font resource dropdown.

This also fixes a number of issues caused by the modelwrapper in the resource
view being changed. This caused a number of issues, because the aliased variables
were still bound to the old modelwrapper, leading to confusion.