Fix inf in Four Point Interpolator Backward
Before this commit, the Four Point Interpolator Backward
wouldn't try hard enough to find a suitable nu, and would end up with
both denominators being 0, resulting in an 'inf' as x.
In Liquify Tool that would cause safe asserts in tile_hash_table
(line 151) when checked for vailidity of the position,
because Grid Interpolation Tools in PaintDevicePolygonOp
(or ImagePolygonOp) didn't check the interpolation point for validity.
And not checking other nu-s would also mean that some results
would be missing.

This commit uses essentially the same logic, except it actually
checks all the values if there is none that is obviously good.
It tries its hardest to exit early but checks everything if it's not
possible.

Tests show that this indeed solves the problem.
It also allows for values slightly outside of the polygon.

This version is slower than before, but it's only slower in
the less common cases. Benchmarks show around 3x slower results.