Rewrite KisWaylandSurfaceColorManager to handle surface recreation by Qt
It seems like Qt has to recreate the wayland surface when
downgrading it from toplevel to subsurface role. We need to
handle that correctly to make sure that we link to the correct
surface all the time. Otherwise we end up being connected
to inexistent surface, so no changes are ever applied to tour
real surface.

This patch adds a sequence of weird initialization states into
KisWaylandSurfaceColorManager. All these states have to be
transitioned before the manages could be considered as
"ready".

Ideally, we should use some kind of state automata framework,
because tryDeinitialize() and tryInitialize() functions are really
hard to understand and maintain.  But I feel like the usage of
QStateMachine would make it even more cryptic.