Make speed input work sensibly for touch drawing
When touch drawing, Krita doesn't receive any hover events when the
finger is in the air, unlike with a mouse or a tablet pen. That means
the usual handling of distance and speed doesn't work, since the inputs
stop where the last stroke left off and start again when the finger is
pressed down again, rather than continually being updated as the pen or
mouse is moving. That leads to the speed input being effectively
garbage at the beginning of a touch stroke.

This commit resets the speed and distance inputs when starting a paint
operation. That doesn't give the same results as with a mouse or pen,
but at least it makes the inputs predictable and useful. The same is
also done when updating the outline so that it doesn't incorrectly
flicker to the wrong size at the beginning of a stroke.

The KoPointerEvent functions to inquire the type of the event are
updated to be const because the outline code passes around a const
event. Not much reason for them not to be const anyway.