Automatically detect touch painting
This changes the touch painting setting from a boolean to a three-state
setting with values auto, enabled and disabled. The latter two work as
they did before. The auto setting will enable touch drawing by default
and then turn it off for the current session when a tablet event is
received. That way, users that have a stylus available will not get the
annoying touch painting blotches when trying to use touch and users that
don't will not be stuck only being able to pan the canvas around, with
neither side having to change the setting manually.

Implements #28.