Fix brush preview outlines being aligned based on incorrect bounding box.
The root cause of this bug is that the class `KisOptimizedBrushOutline` 
does not keep track of the original 'center' of the brush. It instead 
uses the center of the non-transparent pixels in the brush.

The patch just stores the bounds explicitly in `KisOptimizedBrushOutline`
to resolve this issue

BUG:477164 


(cherry picked from commit 036427c473a8243ea96f61200c54de8f180380e7)

46c4af7c Save bounding box of outline for image-based brushes
32b73b36 Fix formatting and add missing include
7ddbc7ab Use const reference for new constructor argument
b980bc3c BUG:477164 Update for review comments

Co-authored-by: Jeff Witthuhn <jlwitthuhn@gmail.com>