Implement loading/saving for text contours.
Because of the particular way svg text shapes handle transforms, we
need to group a textShape and its related shapes if we want to have
the main rotation behaviour rotate both text and shapes at once.

To this end, we write groups in those situations, and add a
`krita:textContourGroup` attribute. This will allow us to recognise
that the group in questions needs to be treated as a textShape with
shapes, while maintaining compatibility with Inkscape.

This also refactors textOnPath to use an ID, which will make it
easier to edit the shapes later, but I am not 100% happy with the
Private being passed just to get the textPath, need to think about this.