[android] Fix main window not showing document
On some Android devices, the main window will sometimes fail to repaint
after a document has loaded. The user can interact with the invisible
contents of the window, but it won't actually show them properly until
Krita is sent to the background once or the user fiddles with the main
menu bar.

Any attempts to call repaint or update don't work, but slightly resizing
the window and back makes it wake up and repaint correctly. This has no
apparent visual effect, since it happens within a single frame, so it
doesn't seem to cause any issues. Devices that don't have the problem to
begin with also don't seem upset about it, unlike the previous solution
of hiding and showing the main window.

BUG:488606