[android] Fix main window not showing document
On some Android devices, the main window will sometimes fail to repaint
after a document has loaded. The user can interact with the invisible
contents of the window, but it won't actually show them properly until
Krita is sent to the background once or the user fiddles with the main
menu bar.

Any attempts to call repaint or update don't work, but hiding and
showing the window on subsequent event loop iterations does, so that's
what this commit does. A QML docker existing also happens to fix this
somehow, so 5.3 with its text properties docker doesn't need this.

BUG:488606