[android] Make IAP and supporter badge work again
This adds a KisAndroidDonations singleton for communication with the
Android-side DonationHelper. The formerly static UI components that got
manually jammed into the interface are now just regularly on the welcome
page and hidden when not on Android, state is communicated through the
singleton via regular signals and slots.

The billing library has been upgraded to version 8. Main thing about
this is that it will automatically reconnect when it loses connection,
so we don't need to deal with that anymore.

The DonationHelper code has been fixed up and somewhat prepared for
additional products. Not too far though, since we haven't nailed down
the details on those yet and that may affect how the code should be
structured. One logical change: pending badge purchases now just show
the badge to the user as if their purchase already succeeded, since we
don't gain anything from holding back the badge except having to write
extra code to handle that case.

Badge purchases are functional, but currently the link to make those is
always disabled, since we don't want to sell those one-time purchase
lifetime badges anymore. Existing lifetime badges are shown to users
that have purchased one in the past.