Change asserts to safe asserts
Before this commit, there were normal asserts in the code
with a recover part that did nothing.
There are two issues with that:
1) It would crash on Android without a chance for the user
to recover,
2) it would be better if it recovered with some reasonable
error.

This commit changes the asserts to safe asserts and changes
the error info to generic error information so the user
is ultimately warned that something went wrong.

Related to #33