Fix color selectors to work in Wayland-HDR mode
This is a huge patch refactoring the HDR management code to support
Wayland's style color management.

1) Adds a special KisMultiSurfaceDisplayConfig structure that strores
a united display surface configuration for both, the GUI and the canvas.

2) Adds KisMultiSurfaceStateManager class that extracts all the surface
format management code into a separate class. It wraps it into a simple
value-based state, which can be transitioned atomically, so we don't have
avalances of partial updates.

3) KisQPainterCanvas and KisOpenGLCanvas2 now do not depend on
KisDisplayColorConverter. Instead, they just receive a proper KisDisplayConfig
on construction and don't care about the rest.

4) All HDR management logic is pulled from KisDisplayColorConverter and
moved to KisMultiSurfaceStateManager. KisDisplayColorConverter now just gets
an already prepared KisMultiSurfaceDisplayConfig object and just distributes
that throughout the rest of Krita.