Implement KisRootSurfaceInfoProxy to get managed surface profile
This proxy object can attach to any widget, track down its toplevel parent,
find its corresponding KisSRGBSurfaceColorSpaceManager and get
information about the currently installed surface mode and profile.

This object is now used in both, KisCanvas2 and KisSmallColorWidget. It
allows not only fetching of the information.

The usage of a special proxy object allows not only fetching of the surface
profile, but also subscription for updates (if they ever happen).