Initial implementation of new text tool options
This switches the old UI file to a QML files that is communicated
with through a Manager object that holds both non-persistent options
as well as persistent options through a Lager model and data system.

Because the tooloptions docker wants to have the tooloptions widget as
small as possible, while the kisquickwidget wants to have the qml
root object follow the widget size, this adds a way to have kisquickwidget
linked to the implicit size of the rootobject.

Beyond that, this exposes toggles for the text cursor so it can switch between
logical and visual mode for bidirectional text, as well as having paste plain
text or paste rich text by default. Next up is allowing selection of a style preset.