Refactor color proofing configuration
The patch implement multiple huge changes:

1) Fixes infinite updates loop in Image Properties dialog when
changing options. The fix is done basically by delegating **all**
the GUI management to lager.

2) Implements KisProofingOptionsWidget, which can be shared between
the image properties dialog and krita preferences dialog.

3) Removes KisProofingConfiguration::adaptationState value for
the absolute colorimetric intent modification (and the corresponding
sliders in the GUI). According to ICCv4 the adaptation should be full
all the time (i.e. value 1.0). And so does Wayland's color management
protocol as of version 1.0.

4) Implements ConversionFlag::NoAdaptationAbsoluteIntent flag which
optionally disables color adaptation to allow the user to prevew the
physical color of the paper without actual adaptation to the display.
This option is not yet implemented in Wayland, but it is planned for
integration:

https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/439

5) Fixes handling of "Save color proofing information to the image" switch. Previously
it was handled by a weird swtich inside KisProofingConfiguration. Now the image just
has **no** proofing configuration when it is not requested by the user. GUI in the image
properties dialog also handles it properly: disabled all the controls when the image
doesn't have its own configuration.

NOTE:
For the backward compatibility reasons, in "file formats", namely in .kra's XML
and in kritarc, adaptation state is still saved as a floating point value of 0.0 or 1.0.
No intermediate values are possible. When loading old files with values between
0.0 and 1.0, the value will be rounded to any side and used as a boolean.