Allow long-pressing to open context menus
By installing an event filter on the application that issues a context
menu event if appropriate. Several types of widgets are excluded by
default since they don't have sensible context menu behaviors to begin
with. Long-press context menu can be explicitly enabled or disabled by
setting a property on the widgets, should that be necessary.

The long-press feature can be enabled or disabled in the preferences. On
Android, it's enabled by default, replacing the previously-existing
right-click emulation on long-press.