Manually define Qt6Gui_PRIVATE_INCLUDE_DIRS CMake variable
Previously Krita was dependent on an implementation detail of Qt's CMake
module, which has a history of breaking. For reference, Krita no longer
builds on Qt dev because they shuffled things around again!

Most notably, the variable Krita used isn't available anymore - so we
need to manually re-create it so the other parts of the CMakeLists
will work. The more correct and reliable solution would be to use the
Qt6Private target itself, but Krita needs to selectively use the private
headers for compilation speed reasons, so I decided to add a workaround
for now.