SAP: Fix eventFilter missing FocusIn events and causing bar to be unmoveable
This is probably a bug in the Canvas first, but this bypasses the selection tool
activating when the mouse focuses "into" the canvas. It doesn't block the ability
for a selection to be activating when changing focus, it only stops it from
preventing the SAP from being notified of a mouse press event.