Implement text on path creation.
Text paths are on a node instead of over the whole text.

This moves the segment location code from KoPathTool to KoPathShape so it can be
reused by the text tool. Hovering over a segment will now show a path icon, where
clicking will create a text on path.

It also implements a number of tests for inserting said path, as well as the ability
to get a toplevel node at the given point.