macos: update krita app icon to display in respect to system appearance
This deprecates the usage of .icns for icon display and uses instead
a generated Assets.car from krita.icon. Now if a user sets a different
appearance than default the krita.app icon will react accordingly.

krita.icns is kept for legacy compatibility but the system ignores it
when running macos<=10.13