macos: working version of thumbnailer
This adds a PreviewHelper app to embed the thumbnailer to it. We need this
as application extensions does not register to lsregister if they are not
embedded inside a bundle application.

Small fixes to all project configurations

The thumbnailer works as expected, special care needs to be taken with
request.scale as it has the current screen scaling, cgImage need to take
this into account.

CCBUG:493474