Replace color sampler preview with a circle
Instead of the rectangle that positions itself perfectly in the way of a
pen being held in the right hand. This circle style is not biased to the
handedness of the user and is used by numerous other programs.

This also gets rid of the rectangle setting since it is unused. If we
want to make the sampler preview configurable, we can reintroduce the
settings accordingly.