Make kinetic scrolling not delay long presses
The long press event filter now checks (with a bit of divination)
whether the click was delayed by kinetic scrolling and if so subtracts
that waited value from the long press time.

This also caps the kinetic scroll click delay to the long press value,
since otherwise your long-presses can get delayed depending on if the
target area is scrollable, which is just weird.