Disable KWin < 6.4.4 workaround  for intent-not-updated bug
KWin < 6.4.4 has a bug resulting in the surface description unchanged
when the only change in surface description is "intent". To workaround
that one should first fully unset the description, commit and then set
it again. This is a potentially crashy workaround, so we shouldn't keep
it active all the time for all compositors.

Ideally, we should detect the version of the compositor and either warn
the user or activate this workaround. But Wayland has no means for
that. I've filed an issue for that:

https://gitlab.freedesktop.org/wayland/wayland-protocols/-/issues/293

Right now the user should manually set environment variable
'KRITA_ENABLE_KWIN_INTENT_WORKAROUND' to activate the
workaround.