Disable long-press on widgets they interfere with
By setting the context menu policy to prevent context menus, which will
make the long-press event filter not consider them as a target. These
widgets don't have context menus anyway, so setting this policy just
makes that explicit and prevents delayed inputs or, in the case of the
popup palette, getting dismissed by the triggered right-click.